Tests to typescript
